+++ crates/miden-lib/src/account/components/mod.rs
@@ -32,14 +32,13 @@ static BASIC_FUNGIBLE_FAUCET_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
     Library::read_from_bytes(bytes).expect("Shipped Basic Fungible Faucet library is well-formed")
 });
 
-// Initialize the Rpo Falcon 512 Procedure ACL library only once.
-static RPO_FALCON_512_PROCEDURE_ACL_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
+// Initialize the Rpo Falcon 512 ACL library only once.
+static RPO_FALCON_512_ACL_LIBRARY: LazyLock<Library> = LazyLock::new(|| {
     let bytes = include_bytes!(concat!(
         env!("OUT_DIR"),
-        "/assets/account_components/rpo_falcon_512_procedure_acl.masl"
+        "/assets/account_components/rpo_falcon_512_acl.masl"
     ));
-    Library::read_from_bytes(bytes)
-        .expect("Shipped Rpo Falcon 512 Procedure ACL library is well-formed")
+    Library::read_from_bytes(bytes).expect("Shipped Rpo Falcon 512 ACL library is well-formed")
 });
 
 // Initialize the NoAuth library only once.
@@ -63,9 +62,9 @@ pub fn rpo_falcon_512_library() -> Library {
     RPO_FALCON_512_LIBRARY.clone()
 }
 
-/// Returns the Rpo Falcon 512 Procedure ACL Library.
-pub fn rpo_falcon_512_procedure_acl_library() -> Library {
-    RPO_FALCON_512_PROCEDURE_ACL_LIBRARY.clone()
+/// Returns the Rpo Falcon 512 ACL Library.
+pub fn rpo_falcon_512_acl_library() -> Library {
+    RPO_FALCON_512_ACL_LIBRARY.clone()
 }
 
 /// Returns the NoAuth Library.
@@ -81,7 +80,7 @@ pub enum WellKnownComponent {
     BasicWallet,
     BasicFungibleFaucet,
     RpoFalcon512,
-    RpoFalcon512ProcedureAcl,
+    RpoFalcon512Acl,
 }
 
 impl WellKnownComponent {
@@ -92,7 +91,7 @@ impl WellKnownComponent {
             Self::BasicWallet => BASIC_WALLET_LIBRARY.mast_forest(),
             Self::BasicFungibleFaucet => BASIC_FUNGIBLE_FAUCET_LIBRARY.mast_forest(),
             Self::RpoFalcon512 => RPO_FALCON_512_LIBRARY.mast_forest(),
-            Self::RpoFalcon512ProcedureAcl => RPO_FALCON_512_PROCEDURE_ACL_LIBRARY.mast_forest(),
+            Self::RpoFalcon512Acl => RPO_FALCON_512_ACL_LIBRARY.mast_forest(),
         };
 
         forest.procedure_digests()
@@ -127,7 +126,7 @@ impl WellKnownComponent {
                     .push(AccountComponentInterface::BasicFungibleFaucet(storage_offset)),
                 Self::RpoFalcon512 => component_interface_vec
                     .push(AccountComponentInterface::AuthRpoFalcon512(storage_offset)),
-                Self::RpoFalcon512ProcedureAcl => component_interface_vec
+                Self::RpoFalcon512Acl => component_interface_vec
                     .push(AccountComponentInterface::AuthRpoFalcon512Acl(storage_offset)),
             }
         }
@@ -142,6 +141,6 @@ impl WellKnownComponent {
         Self::BasicWallet.extract_component(procedures_map, component_interface_vec);
         Self::BasicFungibleFaucet.extract_component(procedures_map, component_interface_vec);
         Self::RpoFalcon512.extract_component(procedures_map, component_interface_vec);
-        Self::RpoFalcon512ProcedureAcl.extract_component(procedures_map, component_interface_vec);
+        Self::RpoFalcon512Acl.extract_component(procedures_map, component_interface_vec);
     }
 }