+++ crates/miden-lib/src/account/interface/component.rs
@@ -56,9 +56,7 @@ impl AccountComponentInterface {
                 "Basic Fungible Faucet".to_string()
             },
             AccountComponentInterface::AuthRpoFalcon512(_) => "RPO Falcon512".to_string(),
-            AccountComponentInterface::AuthRpoFalcon512Acl(_) => {
-                "RPO Falcon512 Procedure ACL".to_string()
-            },
+            AccountComponentInterface::AuthRpoFalcon512Acl(_) => "RPO Falcon512 ACL".to_string(),
             AccountComponentInterface::Custom(proc_info_vec) => {
                 let result = proc_info_vec
                     .iter()