+++ crates/miden-testing/tests/auth/mod.rs
@@ -1,1 +1,1 @@
-mod rpo_falcon_procedure_acl;
+mod rpo_falcon_acl;